Update README.md
If your organization is attempting to shorten the software development process by using AI to generate the code, be aware that there are unseen pitfalls out there.  Some organizations that have tried it have discovered that, while the software creation part of the process make take less time, the software testing process is taking much longer because the engineers, when seeing anomalous behaviors, cannot read the source code to find the source of the behavior.  Essentially, while the code generated by the AI ‘looked good’, it was not ‘human-readable’.  I have developed a process that will allow your development team to collaborate with the AI in a way that the AI acts as a member of the team, and works to ensure that all members of the team understand what it wrote, how the various functions contribute to the final behavior, and how it should be both tested and documented.